+++ calculations/src/dtos/off_peak_tariff.rs
@@ -1,8 +1,8 @@
 #[derive(serde::Deserialize)]
 pub struct CreateDto {
-    pub voltage_level_id: i64,
-    pub power_level_id: i64,
-    pub contract_type_id: i64,
+    pub voltage_level_id: i32,
+    pub power_level_id: i32,
+    pub contract_type_id: i32,
     pub price: f64,
     pub year: u32,
     pub month: u8,
